stub update.

+++ vesin/cuda/src/vesin_cuda_stub.cpp
@@ -4,13 +4,14 @@
 
 using namespace vesin::cuda;
 
+
 void vesin::cuda::free_neighbors(VesinNeighborList & neighbors) {
     throw std::runtime_error("vesin was not compiled with CUDA support");
 }
 
 void vesin::cuda::neighbors(
     const double (*points)[3],
-    size_t n_points,
+    long n_points,
     const double cell[3][3],
     VesinOptions options,
     VesinNeighborList& neighbors
